backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0

SQLAlchemy==2.0.32
psycopg2-binary==2.9.11

python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.4

bcrypt==4.0.1
passlib==1.7.4
python-jose[cryptography]==3.3.0

orjson==3.11.5
reportlab==4.2.2
openpyxl==3.1.5
